+++ packages/terra-form-select/src/single/Frame.jsx
@@ -506,7 +506,6 @@ class Frame extends React.Component {
         aria-label={this.mainLabel()}
         aria-haspopup={!disabled ? 'true' : undefined}
         aria-describedby={ariaDescribedBy}
-        aria-owns={this.state.isOpen ? selectMenuId : undefined}
         aria-required={required}
         className={selectClasses}
         onBlur={this.handleBlur}
@@ -526,15 +525,11 @@ class Frame extends React.Component {
           <span id={descriptionId}>{this.renderDescriptionText()}</span>
         </div>
         {/* Added aria label to avoid announcing empty group by voice over in safari browser */}
-        <div className={cx('display')} aria-label={this.ariaLabel()} aria-hidden={disabled}>
+        <div className={cx('display')} aria-hidden={disabled}>
           {this.getDisplay(displayId, placeholderId)}
         </div>
         {this.renderToggleButton()}
-        <span
-          aria-live={SharedUtil.isSafari() ? 'polite' : 'off'}
-          className={cx('visually-hidden-component')}
-          ref={this.visuallyHiddenComponent}
-        />
+        {this.state.isOpen && <span aria-live={SharedUtil.isSafari() ? 'polite' : 'off'} className={cx('visually-hidden-component')} ref={this.visuallyHiddenComponent} />}
         {this.state.isOpen && (
           <Dropdown
             {...dropdownAttrs}